+++ docs/changelog.md
@@ -4,6 +4,10 @@ title: Changelog
 
 # Changelog
 
+# 3.14.0
+- [`cli`](/helpers/cli/): `validate` cli command to validate a Swagger/OpenAPI file: `npx kubb validate --input swagger.json`
+- [`cli`](/helpers/cli/): `mcp` cli command to start the MCP client to interact with LLMs(like Claude): `npx kubb mcp`
+
 # 3.13.2
 - [`plugin-client`](/plugins/plugin-client): Shadowed variables error when using `client`, use of `fetch` instead when an import to `@kubb/plugin-client/clients/axios` is needed.